crash solved updated

+++ app/src/main/java/com/example/spa/data/response/GraphDataList.kt
@@ -1,7 +1,7 @@
 package com.example.spa.data.response
 
 data class GraphDataList(
-    var amount: Int = 0,
+    var amount: Double = 0.0,
     val date: String = "",
     val value_for: String = ""
 )
